syntax: sh: re-detect shell subtype when shebang changes
If the shebang is changed during editing, and the buffer is reloaded
without being closed, the old subtype variables will stay in place. This
will lead to situations where more than one of these variables is set
and syntax highlighting will be confused.

In my case, I wondered why "select" was not highlighted. Then I changed
the #!/bin/sh to #!/bin/bash, but "select" still stayed unhighlighted
because b:is_posix was still set in addition to b:is_bash.